fixes hw1
