+++ bot/src/test/java/edu/java/bot/commands/TestsMockitoForHelpCommand.java
@@ -1,35 +1,17 @@
 package edu.java.bot.commands;
 
-import com.pengrad.telegrambot.model.Chat;
-import com.pengrad.telegrambot.model.Message;
-import com.pengrad.telegrambot.model.Update;
 import com.pengrad.telegrambot.request.SendMessage;
-import edu.java.bot.commands.HelpCommand;
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-public class TestsForHelpCommand {
-
-    public static Update update;
-    public static Message message;
-    public static Chat chat;
-
+public class TestsMockitoForHelpCommand extends TestsMockitoInitializer {
     @Autowired
     public HelpCommand helpCommand;
-
-    @BeforeAll
-    static void initialize() {
-        update = Mockito.mock(Update.class);
-        message = Mockito.mock(Message.class);
-        chat = Mockito.mock(Chat.class);
-    }
 
     @Test
     @DisplayName("help command handle test")
@@ -38,13 +20,15 @@ public class TestsForHelpCommand {
         Mockito.when(message.text()).thenReturn("/help");
         Mockito.when(message.chat()).thenReturn(chat);
         Mockito.when(chat.id()).thenReturn(11L);
+
         SendMessage sendMessage = helpCommand.handle(update);
+
         Assertions.assertEquals(
             sendMessage.getParameters().get("text"),
             "Available commands:\n" +
+                "/list - Showing the list of tracked links\n" +
                 "/track - Starting tracking a link\n" +
-                "/untrack - Stopping tracking a link\n" +
-                "/list - Showing the list of tracked links\n"
+                "/untrack - Stopping tracking a link\n"
         );
         Assertions.assertEquals(
             sendMessage.getParameters().get("chat_id"),
