+++ bot/src/test/java/edu/java/bot/commands/TestsMockitoForTrackCommand.java
@@ -1,11 +1,7 @@
 package edu.java.bot.commands;
 
-import com.pengrad.telegrambot.model.Chat;
-import com.pengrad.telegrambot.model.Message;
-import com.pengrad.telegrambot.model.Update;
 import com.pengrad.telegrambot.request.SendMessage;
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
@@ -13,29 +9,20 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-public class TestsForUntrackCommand {
-    public static Update update;
-    public static Message message;
-    public static Chat chat;
-
+public class TestsMockitoForTrackCommand extends TestsMockitoInitializer {
     @Autowired
-    public UntrackCommand untrackCommand;
-
-    @BeforeAll
-    static void initialize() {
-        update = Mockito.mock(Update.class);
-        message = Mockito.mock(Message.class);
-        chat = Mockito.mock(Chat.class);
-    }
+    public TrackCommand trackCommand;
 
     @Test
     @DisplayName("track command handle test")
     void handleTest() {
         Mockito.when(update.message()).thenReturn(message);
-        Mockito.when(message.text()).thenReturn("/untrack");
+        Mockito.when(message.text()).thenReturn("/track");
         Mockito.when(message.chat()).thenReturn(chat);
         Mockito.when(chat.id()).thenReturn(11L);
-        SendMessage sendMessage = untrackCommand.handle(update);
+
+        SendMessage sendMessage = trackCommand.handle(update);
+
         Assertions.assertEquals(
             sendMessage.getParameters().get("text"),
             "Currently in development"
