+++ bot/src/test/java/edu/java/bot/commands/TestsMockitoForUntrackCommand.java
@@ -1,11 +1,7 @@
 package edu.java.bot.commands;
 
-import com.pengrad.telegrambot.model.Chat;
-import com.pengrad.telegrambot.model.Message;
-import com.pengrad.telegrambot.model.Update;
 import com.pengrad.telegrambot.request.SendMessage;
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
@@ -13,32 +9,23 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-public class TestsForListCommand {
-    public static Update update;
-    public static Message message;
-    public static Chat chat;
-
+public class TestsMockitoForUntrackCommand extends TestsMockitoInitializer {
     @Autowired
-    public ListCommand listCommand;
-
-    @BeforeAll
-    static void initialize() {
-        update = Mockito.mock(Update.class);
-        message = Mockito.mock(Message.class);
-        chat = Mockito.mock(Chat.class);
-    }
+    public UntrackCommand untrackCommand;
 
     @Test
-    @DisplayName("list command handle test")
+    @DisplayName("track command handle test")
     void handleTest() {
         Mockito.when(update.message()).thenReturn(message);
-        Mockito.when(message.text()).thenReturn("/list");
+        Mockito.when(message.text()).thenReturn("/untrack");
         Mockito.when(message.chat()).thenReturn(chat);
         Mockito.when(chat.id()).thenReturn(11L);
-        SendMessage sendMessage = listCommand.handle(update);
+
+        SendMessage sendMessage = untrackCommand.handle(update);
+
         Assertions.assertEquals(
             sendMessage.getParameters().get("text"),
-            "No links are being tracked currently"
+            "Currently in development"
         );
         Assertions.assertEquals(
             sendMessage.getParameters().get("chat_id"),
